fix: resolve Next.js route export build error and add missing E2E dashboard widget mocks
